==> 01_coupling_NREM_CFS.ipynb <==
# NSRR CFS Features extraction

https://sleepdata.org/datasets/cfs/pages/equipment/montage-and-sampling-rate-information.md

import os
import yasa
import warnings
import numpy as np
import pandas as pd
import pingouin as pg
from tqdm.notebook import tqdm
from mne.io import read_raw_edf
from scipy.signal import hilbert
from mne.filter import filter_data

from tensorpac import methods as tpm
from tensorpac import EventRelatedPac

import seaborn as sns
import matplotlib.pyplot as plt
sns.set(font_scale=1.25)

from helper_functions import get_sub_visit_hyp

# Define paths
root_dir = '/Volumes/NSRR/cfs/'
eeg_dir = root_dir + 'polysomnography/edfs/'
hypno_dir = root_dir + 'polysomnography/annotations-events-profusion/'

# List all EDF files
all_edfs = os.listdir(eeg_dir)
all_edfs = all_edfs = [c for c in os.listdir(eeg_dir) if c.endswith(".edf")]
all_subj = [c.split("-")[2][:-4] for c in all_edfs]

print(len(all_subj))

df = []
df_erpac = []
include = ['C3', 'M2']
sf = 100

np.seterr(all="ignore")

for i, sub in enumerate(tqdm(all_subj)):
    
    eeg_file = eeg_dir + 'cfs-visit5-' + str(sub) + '.edf'
    hypno_file = hypno_dir + 'cfs-visit5-' + str(sub) + '-profusion.xml'
    
    # Check that file exists
    if not os.path.isfile(eeg_file):
        warnings.warn("File not found %s" % eeg_file)
        continue
    if not os.path.isfile(hypno_file):
        warnings.warn("File not found %s" % hypno_file)
        continue

    # LOAD EEG DATA
    try:
        raw = read_raw_edf(eeg_file, preload=False, verbose=0)
        raw.drop_channels(np.setdiff1d(raw.info['ch_names'], include))
        raw.load_data()
    except:
        continue
        
    # Resample to 100 Hz and low-pass filter 
    raw.resample(sf, npad="auto")
    
    # Re-reference to C3-M2
    raw.set_eeg_reference(["M2"])
    raw.drop_channels(['M2'])

    # LOAD HYPNOGRAM
    hypno, sf_hyp = yasa.load_profusion_hypno(hypno_file)
    # Check that hypno and data have the same number of epochs
    n_epochs = hypno.shape[0]
    if n_epochs != np.floor(raw.n_times / sf / 30):
        warnings.warn("Hypno and data size do not match. Skipping subject.")
        continue
        
    # Set NREM is 6
    hypno_NREM = pd.Series(hypno).replace({2: 6, 3: 6}).to_numpy()
        
    # Upsample hypno NREM
    hypno_NREM_up = yasa.hypno_upsample_to_data(hypno_NREM, sf_hyp, raw)
    
    # Extract C3 data
    data = np.squeeze(raw.get_data(units="uV"))
    assert data.ndim == 1
    
    # Invert polarity
    # http://zzz.bwh.harvard.edu/luna/vignettes/nsrr-polarity/
    data *= -1
    
    #################################
    # PHASE-AMPLITUDE COUPLING
    #################################
    
    # Slow-waves detection in NREM (N2 + N3)
    ptp_thr = 60
    neg_amp = round(ptp_thr / 1.875)
    freq_sw = (0.3, 1.5)
    freq_sp = (12, 16)
    sw = yasa.sw_detect(
            data, sf, ch_names=["EEG"], hypno=hypno_NREM_up, include=6, freq_sw=freq_sw,
            dur_neg=(0.3, 1.5), dur_pos=(0.1, 1), amp_neg=(neg_amp, 200),
            amp_pos=(10, 200), amp_ptp=(ptp_thr, 300), coupling=False,
            freq_sp=freq_sp, remove_outliers=True
    )
    
    if sw is None:
        warnings.warn("No SOs detected for %s. SKIPPING SUBJECT." % sub)
        continue
        
    # Spindles detection in NREM (N2 + N3)
    sp = yasa.spindles_detect(
            data, sf, ch_names=["EEG"], hypno=hypno_NREM_up, include=6, duration=(0.4, 2), 
            thresh={'rel_pow': 0.05, 'corr': 0.50, 'rms': 2}, remove_outliers=True)
    
    if sp is None:
        warnings.warn("No spindles detected for %s. SKIPPING SUBJECT." % sub)
        continue
        
    sp_sum = sp.summary(grp_stage=True)
        
    # Co-occurring slow-waves / spindles and get summary
    sw_sum = sw.find_cooccurring_spindles(sp.summary())
    sw_sum = sw.summary(grp_chan=False, grp_stage=True)
    
    # Assert that we have at least 5 slow-waves, otherwise skip subject
    if sw._events.shape[0] <= 5:
        warnings.warn("Less than 5 SOs detected for %s. SKIPPING SUBJECT." % sub)
        continue
    
    # PAC using only +/- 1 second around negative peak, as in Mikutta et al 2021 (J. Sleep Research)
    data_sp = filter_data(
        sw._data, sf, 12, 16, method='fir', l_trans_bandwidth=1.5,
        h_trans_bandwidth=1.5, verbose=0)
    amp = np.squeeze(np.abs(hilbert(data_sp)))
    pha = np.squeeze(np.angle(hilbert(sw._data_filt)))
    
    # Get epochs
    idx_pks = (sw._events['NegPeak'] * sf).to_numpy().astype(int)  # In samples
    idx_epochs = yasa.get_centered_indices(amp, idx_pks, npts_before=sf, npts_after=sf)[0]
    
    amp = amp[idx_epochs][None, ...]  # shape (n_freqs, n_epochs, n_samples)
    pha = pha[idx_epochs][None, ...]
    
    # Calculate raw PAC and preferred phase
    ndp = np.squeeze(tpm.norm_direct_pac(pha, amp, p=1))
    ndp_thr = np.squeeze(tpm.norm_direct_pac(pha, amp))  # Unreliable values are set to 0
    idx_thr_supzero = np.nonzero(ndp_thr)
    ndp_thr_supzero = ndp_thr[idx_thr_supzero]
    pp = np.squeeze(tpm.preferred_phase(pha, amp, n_bins=18)[1])
    pp_thr_supzero = pp[idx_thr_supzero]
    
    # Append to output dict
    output_dict = {
        "dataset": "cfs",
        "subj": sub,
        # Spindles
        "sp_density": sp_sum.loc[6, "Density"],
        "sp_power": sp_sum.loc[6, "RelPower"],
        "sp_freq": sp_sum.loc[6, "Frequency"],
        # Slow-waves
        "sw_density": sw_sum.loc[6, "Density"],
        "sw_ptp": sw_sum.loc[6, "PTP"],
        "sw_freq": sw_sum.loc[6, "Frequency"],
        "sw_ndpac": np.nanmean(ndp),
        "sw_ndpac_thr": np.nanmean(ndp_thr),
        "sw_ndpac_thr_supzero": np.nanmean(ndp_thr_supzero),
        "sw_ndpac_prop_supzero": ndp_thr_supzero.size / ndp_thr.size,
        "sw_pp": pg.circ_mean(pp),
        "sw_pp_thr_supzero": pg.circ_mean(pp_thr_supzero),
        "sw_coinc_spindles": sw_sum.loc[6, "CooccurringSpindle"],
    }

    # Append to main dataframe
    df.append(output_dict)
    
    #################################
    # ERPAC
    #################################
    
    # Calculate event-related PAC
    # Get peak-locked data +/- 3 seconds before negative peak of each slow-waves
    # Events that are outside the bound of data are masked in "idx_epochs"
    idx_epochs, idx_valid = yasa.get_centered_indices(
        data, idx_pks, npts_before=int(3 * sf), npts_after=int(3 * sf))
    
    if idx_valid.size != idx_pks.size:
        warnings.warn("Some events-locked windows are outside the edge of data. Skipping subject %s for ERPAC." % sub)
        continue
        
    # Extract an example of a coupled SO and an uncoupled one
    if i == 0:
        df_so_example = pd.DataFrame({
            "sp_coupling": data_sp[0, idx_epochs[600]][200:401],
            "so_coupling": sw._data_filt[0, idx_epochs[600]][200:401],
            "sp_nocoup": data_sp[0, idx_epochs[269]][200:401],
            "so_nocoup": sw._data_filt[0, idx_epochs[269]][200:401]})
        
        df_so_example.to_csv("output/csv/df_example_so.csv", index=False)

    # Keep only events with ndPAC>0
    idx_epochs = idx_epochs[idx_thr_supzero]

    # Average slow-waves
    avg_sw = np.squeeze(sw._data_filt[:, idx_epochs]).mean(0)
    time = np.arange(len(avg_sw)) / sf
    
    df_avg_sw = pd.DataFrame({
        "subj": sub,
        "time": time,
        "avg_sw": avg_sw,
    })
    
    # ERPAC (+/- 3 sec to avoid filter edge)
    data_erpac = data[idx_epochs] 
    erp = EventRelatedPac(f_pha=[0.3, 1.5], f_amp=np.arange(4.75, 25.75, 0.5), verbose=False)  # f_pha = 0.8 Hz
    freqs = erp.f_amp.mean(1).astype(str)
    pha = erp.filter(sf, data_erpac, ftype='phase')
    amp = erp.filter(sf, data_erpac, ftype='amplitude')
    erpac = np.squeeze(erp.fit(pha, amp, method="circular"))  # Fast
    ergcpac = np.squeeze(erp.fit(pha, amp, method="gc", smooth=50))  # Slow -- Gaussian Copula
    
    # Merge into a single dataframe
    erpac = pd.DataFrame(erpac.T)
    erpac.columns = freqs
    erpac = erpac.add_prefix("circ_")
    erpac['time'] = time
    
    ergcpac = pd.DataFrame(ergcpac.T)
    ergcpac.columns = freqs
    ergcpac = ergcpac.add_prefix("gc_")
    ergcpac['time'] = time
    
    df_erpac.append(df_avg_sw.merge(erpac, how="inner").merge(ergcpac, how="inner"))
    
    # Save ERPAC dataframe every 10 subjects
    if i % 20 == 0:
        pd.concat(df_erpac).set_index(['subj', 'time']).to_parquet("output/csv/df_cfs_ERGCPAC_inverted.parquet")

df = pd.DataFrame(df).set_index(["dataset", "subj"])
df.round(3)

df.to_csv("output/csv/df_cfs_coupling_NREM_inverted.csv")

# Save ER(GC)PAC
pd.concat(df_erpac).set_index(['subj', 'time']).to_parquet("output/csv/df_cfs_ERGCPAC_inverted.parquet")
